Speed up pathfinder by decreasing accuracy
The hybrid A* algorithm's default delta angle at the goal is 6 degrees, meaning the goal
node must be within this angle for the goal to be considered as reached. This may lead
to very long or completely unsuccessful path finding if the goal is close to an obstacle
for example.
We'll therefore start relaxing this criteria so bigger angle differences are acceptable
as the pathfinder number of unsuccessful iterations grow.